packagedb/requirements.txt
asgiref==3.5.2
Django==4.1.2
django-filter==22.1
djangorestframework==3.14.0
natsort==8.2.0
packageurl-python==0.10.4
psycopg2-binary==2.9.3
pytz==2022.5
setuptools==65.3.0
sqlparse==0.4.3
